feat: Starting the Kubernetes adjustments
