+++ ant403/sample-apps/00-fluentBit/fluent-bit.conf
@@ -1,0 +1,21 @@
+[SERVICE]
+    Flush        5
+    Daemon       Off
+    Log_Level    debug
+
+[INPUT]
+    Name   forward
+    Listen 0.0.0.0
+    Port   24224
+
+[OUTPUT]
+    Name  es
+    Match *
+    Host  OSS_DOMAIN
+    Port  443
+    Index sample_app_logs
+    Type  docker
+    HTTP_User OSSDOMAIN_USERNAME
+    HTTP_Passwd OSSDOMAIN_PASSWORD
+    tls On
+    tls.verify Off
