+++ ant403/sample-apps/01-data-preper/kubernetes/00-namespace.yaml
@@ -1,0 +1,4 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: data-preper
