+++ ant403/sample-apps/01-data-preper/kubernetes/01-configmap.yaml
@@ -1,0 +1,80 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: data-preper-config
+  namespace: data-preper
+data:
+  root-ca.pem: |
+    -----BEGIN CERTIFICATE-----
+    MIID/jCCAuagAwIBAgIBATANBgkqhkiG9w0BAQsFADCBjzETMBEGCgmSJomT8ixk
+    ARkWA2NvbTEXMBUGCgmSJomT8ixkARkWB2V4YW1wbGUxGTAXBgNVBAoMEEV4YW1w
+    bGUgQ29tIEluYy4xITAfBgNVBAsMGEV4YW1wbGUgQ29tIEluYy4gUm9vdCBDQTEh
+    MB8GA1UEAwwYRXhhbXBsZSBDb20gSW5jLiBSb290IENBMB4XDTE4MDQyMjAzNDM0
+    NloXDTI4MDQxOTAzNDM0NlowgY8xEzARBgoJkiaJk/IsZAEZFgNjb20xFzAVBgoJ
+    kiaJk/IsZAEZFgdleGFtcGxlMRkwFwYDVQQKDBBFeGFtcGxlIENvbSBJbmMuMSEw
+    HwYDVQQLDBhFeGFtcGxlIENvbSBJbmMuIFJvb3QgQ0ExITAfBgNVBAMMGEV4YW1w
+    bGUgQ29tIEluYy4gUm9vdCBDQTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoC
+    ggEBAK/u+GARP5innhpXK0c0q7s1Su1VTEaIgmZr8VWI6S8amf5cU3ktV7WT9SuV
+    TsAm2i2A5P+Ctw7iZkfnHWlsC3HhPUcd6mvzGZ4moxnamM7r+a9otRp3owYoGStX
+    ylVTQusAjbq9do8CMV4hcBTepCd+0w0v4h6UlXU8xjhj1xeUIz4DKbRgf36q0rv4
+    VIX46X72rMJSETKOSxuwLkov1ZOVbfSlPaygXIxqsHVlj1iMkYRbQmaTib6XWHKf
+    MibDaqDejOhukkCjzpptGZOPFQ8002UtTTNv1TiaKxkjMQJNwz6jfZ53ws3fh1I0
+    RWT6WfM4oeFRFnyFRmc4uYTUgAkCAwEAAaNjMGEwDwYDVR0TAQH/BAUwAwEB/zAf
+    BgNVHSMEGDAWgBSSNQzgDx4rRfZNOfN7X6LmEpdAczAdBgNVHQ4EFgQUkjUM4A8e
+    K0X2TTnze1+i5hKXQHMwDgYDVR0PAQH/BAQDAgGGMA0GCSqGSIb3DQEBCwUAA4IB
+    AQBoQHvwsR34hGO2m8qVR9nQ5Klo5HYPyd6ySKNcT36OZ4AQfaCGsk+SecTi35QF
+    RHL3g2qffED4tKR0RBNGQSgiLavmHGCh3YpDupKq2xhhEeS9oBmQzxanFwWFod4T
+    nnsG2cCejyR9WXoRzHisw0KJWeuNlwjUdJY0xnn16srm1zL/M/f0PvCyh9HU1mF1
+    ivnOSqbDD2Z7JSGyckgKad1Omsg/rr5XYtCeyJeXUPcmpeX6erWJJNTUh6yWC/hY
+    G/dFC4xrJhfXwz6Z0ytUygJO32bJG4Np2iGAwvvgI9EfxzEv/KP+FGrJOvQJAq4/
+    BU36ZAa80W/8TBnqZTkNnqZV
+    -----END CERTIFICATE-----
+  data-prepper-config.yaml: | 
+    ssl: false
+  data-prepper-wait-for-odfe-and-start.sh: |
+    #!/bin/bash
+
+    until [[ $(curl --write-out %{http_code} --output /dev/null --silent --head --fail https://OSS_DOMAIN:443 -u OSSDOMAIN_USERNAME:OSSDOMAIN_PASSWORD --insecure) == 200 ]]; do
+      echo "Waiting for OpenSearch to be ready"
+      sleep 1
+    done
+    java -Xms128m -Xmx128m -jar data-prepper.jar /usr/share/data-prepper/pipelines.yaml /usr/share/data-prepper/data-prepper-config.yaml
+  
+  pipelines.yaml: |
+    entry-pipeline:
+      delay: "100"
+      source:
+        otel_trace_source:
+          ssl: false
+      sink:
+        - pipeline:
+            name: "raw-pipeline"
+        - pipeline:
+            name: "service-map-pipeline"
+    raw-pipeline:
+      source:
+        pipeline:
+          name: "entry-pipeline"
+      prepper:
+        - otel_trace_raw_prepper:
+      sink:
+        - opensearch:
+            hosts: [ "https://vpc-aosstraceanalytics-iimypnfcvmlcyzcmjo7yzn34rq.us-east-1.es.amazonaws.com" ]
+            username: "aosadmin"
+            password: "Passw0rd1!"
+            trace_analytics_raw: true
+    service-map-pipeline:
+      delay: "100"
+      source:
+        pipeline:
+          name: "entry-pipeline"
+      prepper:
+        - service_map_stateful:
+      sink:
+        - opensearch:
+            hosts: [ "https://vpc-aosstraceanalytics-iimypnfcvmlcyzcmjo7yzn34rq.us-east-1.es.amazonaws.com" ]
+            username: "aosadmin"
+            password: "Passw0rd1!"
+            trace_analytics_service_map: true
+
+
