+++ ant403/sample-apps/01-data-preper/kubernetes/02-deployment.yaml
@@ -1,0 +1,55 @@
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  labels:
+    app: data-preper
+  name: data-preper
+  namespace: data-preper
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: data-preper
+  template:
+    metadata:
+      labels:
+        app: data-preper
+    spec:
+      containers:
+      - image: opensearchproject/data-prepper:latest
+        name: data-preper
+        ports:
+        - containerPort: 21890
+        resources: # Verify the amount of requests
+          requests:
+            memory: "128Mi"
+            cpu: "250m"
+          limits:
+            memory: "256Mi"
+            cpu: "500m"
+        command: ["sh", "-c", "java -Xms128m -Xmx128m -jar data-prepper.jar /usr/share/data-prepper/pipelines.yaml /usr/share/data-prepper/data-prepper-config.yaml"]
+        volumeMounts:
+        - name: config-volume
+          mountPath: /usr/share/data-prepper/data-prepper-wait-for-odfe-and-start.sh # Do not use this script
+          subPath: data-prepper-wait-for-odfe-and-start.sh
+        - name: config-volume
+          mountPath: /usr/share/data-prepper/pipelines.yaml
+          subPath: pipelines.yaml
+        - name: config-volume
+          mountPath: /usr/share/data-prepper/data-prepper-config.yaml
+          subPath: data-prepper-config.yaml
+        - name: config-volume
+          mountPath: /usr/share/data-prepper/root-ca.pem
+          subPath: root-ca.pem
+        env:
+        - name: OTLP
+          value: ""
+        - name: MYSQL_HOST
+          value: "mysql.mysql.svc.cluster.local"
+        - name: MYSQL_PORT
+          value: "3306"
+      volumes:
+      - name: config-volume
+        configMap:
+          name: data-preper-config
