+++ ant403/sample-apps/01-data-preper/kubernetes/03-service.yaml
@@ -1,0 +1,12 @@
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: data-preper
+  namespace: data-preper
+spec:
+  ports:
+  - port: 80
+    targetPort: 21890
+  selector:
+    app: data-preper
