+++ ant403/sample-apps/02-otel-collector/kubernetes/00-namespace.yaml
@@ -1,0 +1,4 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: otel-collector
