+++ ant403/sample-apps/02-otel-collector/kubernetes/01-configmap.yaml
@@ -1,0 +1,47 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: otel-collector-config
+  namespace: otel-collector
+data:
+  otel-collector-config.yml: |
+    receivers:
+      otlp:
+        protocols:
+          grpc:
+            endpoint: 0.0.0.0:55680
+            
+    exporters:
+      otlp/2:
+        endpoint: data-preper.data-preper.svc.cluster.local:80
+        tls:
+          insecure: true
+      
+
+    service:
+      pipelines:
+        traces:
+          receivers: [otlp]
+          exporters: [otlp/2]
+
+  demo-data-prepper.crt: | 
+    -----BEGIN CERTIFICATE-----
+    MIICqjCCAZICCQDBYJxuuzsbczANBgkqhkiG9w0BAQsFADAXMRUwEwYDVQQDDAxk
+    YXRhLXByZXBwZXIwHhcNMjAxMTMwMTc0NDM4WhcNMzAxMTI4MTc0NDM4WjAXMRUw
+    EwYDVQQDDAxkYXRhLXByZXBwZXIwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEK
+    AoIBAQDBuSTI3oz9ZJRiVQCdTfbQlG9TWCXwuIjjJcVwvgVjhIuj5VLEOWUpC8nr
+    be4/3ttGLuqOBViLlw2aelN7ZANgFy33ONF7kDh4xFMp4rZdFKxCOsmUpbSD7hRf
+    8mVRFv30bx9//bVjLP8qi2QKuHc8WOevMsdNENFI6BgiI52EoVPyYJIIMaE2ripG
+    qkQVn04Tk+rWz3zQ+NyOcExow98+pxMIjAanMUELTpMjQCWq2HtV2nnqcfqqNCae
+    ig7NSDRvQtPL84Fjy7y4B1pS558hBmfOyZcGHo/gxLxQmvb1UKy4bcL7Kd8H3aNr
+    LZ8zsr+he3cM9a7SZbiilALdcv0NAgMBAAEwDQYJKoZIhvcNAQELBQADggEBAKvw
+    fLiflX0Wd54ExsQvLvwZbrVU0z110h2N+E7SXt9K4NhHsANsX2mWxbqU/0YMYQpo
+    fPO0XJSqmo28YD/eyKEazJg7tFISEqQ5gfWZM8BkfHcjbHMixtRDm6SpNzfeQT97
+    C0zZo5kAD4d0Y3FwVjQ5FkvlmM6rCaJ0NyvZBUJseCsWpZsV2vF0QbMQcQ0lxjqY
+    Vxs8Cnwp9bfF7ZP3i5KD5JxtVm80g6MIQdoYDQqtMjhZ8rCg5e1S6DbPnLRlmXkz
+    4cFfD4V/7+ldSpEt38QBtCFWalfem+4dtbEX8FoTciNZfCkeUotpO2jPS1IzzghA
+    yy6bLJd2MHl8AnGRdQg=
+    -----END CERTIFICATE-----
+  
+
+
