+++ ant403/sample-apps/02-otel-collector/kubernetes/03-service.yaml
@@ -1,0 +1,12 @@
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: otel-collector
+  namespace: otel-collector
+spec:
+  ports:
+  - port: 80
+    targetPort: 55680
+  selector:
+    app: otel-collector
