+++ ant403/sample-apps/04-analytics-service/settings.gradle
