+++ ant403/sample-apps/04-analytics-service/src/main/resources/application.properties
