+++ ant403/sample-apps/05-databaseService/kubernetes/00-namespace.yaml
@@ -1,0 +1,5 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: database-service
+
