+++ ant403/sample-apps/06-orderService/requirements.txt
@@ -1,0 +1,10 @@
+dash==1.17.0
+mysql-connector==2.2.9
+opentelemetry-exporter-otlp==1.9.1
+opentelemetry-instrumentation-flask==0.28b1
+opentelemetry-instrumentation-mysql==0.28b1
+opentelemetry-instrumentation-requests==0.28b1
+opentelemetry-instrumentation-logging==0.28b1
+opentelemetry-sdk==1.9.1
+flask==2.0.2
+protobuf==3.15.6
