+++ ant403/sample-apps/08-paymentService/paymentService.py
@@ -1,6 +1,6 @@
-# AuthenticationService running on port 8085
+# PaymentService running on port 8084
 
-from flask import Flask, jsonify, request
+from flask import Flask, jsonify, request, make_response
 import logging
 from opentelemetry import trace
 from opentelemetry.instrumentation.logging import LoggingInstrumentor
@@ -16,23 +16,23 @@ from opentelemetry.sdk.trace.export import (
 from Error import Error
 from requests.adapters import HTTPAdapter
 from urllib3.util.retry import Retry
-import flask, os, pkg_resources, socket, random, requests, json
+import os, pkg_resources, socket, requests, random, json
 
 logger = logging.getLogger(__name__)
 
-LOGIN_ERROR_RATE_THRESHOLD = 0
+ERROR_RATE_THRESHOLD = 60
 
 app = Flask(__name__)
 
 OTLP = os.getenv("OTLP") if os.getenv("OTLP") is not None else "localhost"
-RECOMMEND = os.getenv("RECOMMEND") if os.getenv("RECOMMEND") is not None else "localhost"
+INVENTORY = os.getenv("INVENTORY") if os.getenv("INVENTORY") is not None else "localhost"
 LOGS = os.getenv("LOGS") if os.getenv("LOGS") is not None else "localhost"
 
 trace.set_tracer_provider(
     TracerProvider(
         resource=Resource.create(
             {
-                "service.name": "authentication",
+                "service.name": "payment",
                 "service.instance.id": str(id(app)),
                 "telemetry.sdk.name": "opentelemetry",
                 "telemetry.sdk.language": "python",
@@ -69,27 +69,36 @@ def handle_invalid_usage(error):
     response.status_code = error.status_code
     return response
 
-@app.route("/server_request_login")
-def server_request_login():
+@app.route("/checkout", methods=["POST", "GET"])
+def payment():
     errorRate = random.randint(0,99)
-    with tracer.start_as_current_span("verify_login"):
-        if errorRate < LOGIN_ERROR_RATE_THRESHOLD:
-            logs('Authentication', 'Customer failed login - Unauthenticated: 401')
-            logger.error('Authentication - Customer failed login - Unauthenticated: 401')
-            raise Error('Failed Login', status_code=401.1)
+    if errorRate < ERROR_RATE_THRESHOLD:
+        logs('Payment', 'Checkout operation failed - Service Unavailable: 503')
+        logger.error('Payment - Checkout operation failed - Service Unavailable: 503')
+        raise Error('Checkout Failed - Service Unavailable', status_code=503)
+    else:
+        with tracer.start_as_current_span("checkout"):
+            rawData = request.form
+            data = {}
+            for itemId in rawData.keys():
+                data[itemId] = sum([-val for val in rawData.getlist(itemId, type=int)])
 
-    # Successfully logged in @ this point, return product recommendations to client.
-    recommendationSession = requests.Session()
-    recommendationSession.mount("http://", HTTPAdapter(max_retries=retry_strategy))
-    getRecommendationAPIResponse = recommendationSession.get(
-        "http://{}:8086/recommend".format(RECOMMEND)
-    )
-    recommendationSession.close()
-
-    assert getRecommendationAPIResponse.status_code == 200
-    logs('Authentication', 'Customer successful login')
-    logger.info('Authentication - Customer successful login')
-    return getRecommendationAPIResponse.json()
+            soldInventorySession = requests.Session()
+            soldInventorySession.mount("http://", HTTPAdapter(max_retries=retry_strategy))
+            soldInventoryUpdateResponse = soldInventorySession.post(
+                "http://{}:8082/update_inventory".format(INVENTORY),
+                data=data,
+            )
+            soldInventorySession.close()
+            if soldInventoryUpdateResponse.status_code == 200:
+                logs('Payment', 'Customer successfully checked out cart')
+                logger.info('Payment - Customer successfully checked out cart')
+                return "success"
+            else:
+                failedItems = soldInventoryUpdateResponse.json().get("failed_items")
+                return make_response(
+                    "Failed to checkout following items: {}".format(','.join(failedItems)),
+                    soldInventoryUpdateResponse.status_code)
 
 def logs(serv=None, mes=None):
     create_log_data = {'service': serv, 'message': mes}
@@ -102,4 +111,4 @@ def logs(serv=None, mes=None):
     return "success"
 
 if __name__ == "__main__":
-    app.run(port=8085, host="0.0.0.0")
+    app.run(port=5000, host="0.0.0.0")
